Power Bi Road Accident Analysis Dashboard
I created this dashboard using a dummy data of the United Kingdom to break down the number of road accidents based on location, vehicle types, road types, time of the day, current year versus the previous year to see if there has been an increase or decrease in the road accidents over a period of one year. In order to build the dashboard and visualizations, I had to build a date table and then set up a one to many relationship with the accident date column from the imported data. I also cleaned the data using the power query editor and this made the data a lot easier to understand and use to build visualizations.